pyproject.toml
[tool.poetry]
name = "python-playwright-test-framework"
version = "0.1.0"
description = ""
authors = ["m.biletskyi"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
playwright = "^1.48.0"
pytest-playwright = "^0.5.2"
python-dotenv = "^1.0.1"
pytest-xdist = "^3.6.1"
pytest-reporter-html1 = "^0.9.2"
pytest-sugar = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
pylint = "^3.3.1"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"